+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.0
+
+- Changed `numbersToEnglish` and `numbersToPersian` to String extension functions
+- Changed `gregorianToJalali` and `jalaliToGregorian` to static functions
+- Fixed minor bugs
+- Improved code clarity
+
+
 ## 0.1.0
 
 - Initial version